refactor: publish the core package as @verrex/core
The unscoped `verrex` name is blocked by npm's typosquat-similarity
guard (too close to `verror`). Scope the core package under the org
instead: `verrex` → `@verrex/core`.

- package name + the demo/ts-plugin workspace deps
- every `from "verrex…"` import specifier + the compiler's emitted
  runtime import (`RUNTIME_PKG`) → `@verrex/core…`
- self-referencing subpath imports resolve via the renamed package's
  own exports map
- escaped the `/` in the compiler tests' regex assertions

The framework is still "verrex" everywhere else (the `.vx` extension,
`VerrexLive`, the `packages/verrex` dir, the `verrex-check` bin).
`@verrex/ts-plugin` is unaffected (already scoped, already published).
