+++ apps/demo/src/LiveUser.vx
@@ -1,6 +1,6 @@
 import { Cause, Effect } from "effect"
 import { AtomRef } from "effect/unstable/reactivity"
-import { Await } from "verrex"
+import { Await } from "@verrex/core"
 import { Http } from "./services.ts"
 
 /**
