+++ packages/ts-plugin/package.json
@@ -27,7 +27,7 @@
     "typescript": "^6.0.3"
   },
   "devDependencies": {
-    "verrex": "workspace:*",
+    "@verrex/core": "workspace:*",
     "@volar/language-core": "^2.4.28",
     "@volar/typescript": "^2.4.28",
     "@effect/vitest": "4.0.0-beta.71",
