+++ packages/verrex/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "verrex",
+  "name": "@verrex/core",
   "version": "0.1.0",
   "description": "An Effect-native UI framework where Effect's <A, E, R> channels propagate from every leaf of the view tree to the root — a forgotten Layer becomes a compile-time error that names the missing service.",
   "license": "MIT",
