+++ packages/verrex/src/check/index.ts
@@ -3,7 +3,7 @@ import * as ts from "typescript"
 import * as kit from "@volar/kit"
 import { create as createTypeScriptServices } from "volar-service-typescript"
 import { URI } from "vscode-uri"
-import { createVerrexLanguagePlugin } from "verrex/language"
+import { createVerrexLanguagePlugin } from "@verrex/core/language"
 
 // LSP DiagnosticSeverity constants (stable wire protocol). `@volar/language-service`
 // re-exports the type but not the values, so we inline these to avoid pulling in
