+++ packages/verrex/src/language/index.ts
@@ -1,4 +1,4 @@
 export { createVerrexLanguagePlugin, type TypeScriptConfig } from "./language-plugin.ts"
 export { convertSourceMap } from "./source-map.ts"
 export { VerrexVirtualCode } from "./virtual-code.ts"
-export type { JsxRange } from "verrex/compiler"
+export type { JsxRange } from "@verrex/core/compiler"
